add date jalali picker & role validate jalali

+++ app/Http/Livewire/ShowTasks.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire;
 
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 use Livewire\Component;
 
 use Livewire\WithPagination;
@@ -12,9 +13,10 @@ use App\Models\Tasks;
 
 class ShowTasks extends Component
 {
-
+    //for pagination Live wire
     use WithPagination;
 
+    //template pagination
     protected $paginationTheme = 'bootstrap';
 
     //for filter object
@@ -24,11 +26,15 @@ class ShowTasks extends Component
     public $order_by ='id';
     public $order = 'asc';//desc
 
-    //icon sorting
-    public $order_icon = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-sort-down-alt" viewBox="0 0 16 16">
-  <path d="M3.5 3.5a.5.5 0 0 0-1 0v8.793l-1.146-1.147a.5.5 0 0 0-.708.708l2 1.999.007.007a.497.497 0 0 0 .7-.006l2-2a.5.5 0 0 0-.707-.708L3.5 12.293V3.5zm4 .5a.5.5 0 0 1 0-1h1a.5.5 0 0 1 0 1h-1zm0 3a.5.5 0 0 1 0-1h3a.5.5 0 0 1 0 1h-3zm0 3a.5.5 0 0 1 0-1h5a.5.5 0 0 1 0 1h-5zM7 12.5a.5.5 0 0 0 .5.5h7a.5.5 0 0 0 0-1h-7a.5.5 0 0 0-.5.5z"/>
-</svg>';//desc
+    //message alert
+    public $message_type,$message='';
 
+    public $users;
+    public $errors_message=[];
+
+    public function mount(){
+        $this->users = User::all();//get list user for tasks user id & create user filter
+    }
 
     /**
      * change order by & order click header table
@@ -37,15 +43,28 @@ class ShowTasks extends Component
     public function orderBy($field_name){
         $this->order_by = $field_name;
         $this->order = $this->order == 'asc' ? 'desc' :'asc';
-        $this->order_icon = $this->order == 'desc' ? '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-sort-up" viewBox="0 0 16 16">
-  <path d="M3.5 12.5a.5.5 0 0 1-1 0V3.707L1.354 4.854a.5.5 0 1 1-.708-.708l2-1.999.007-.007a.498.498 0 0 1 .7.006l2 2a.5.5 0 1 1-.707.708L3.5 3.707V12.5zm3.5-9a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 0 1h-7a.5.5 0 0 1-.5-.5zM7.5 6a.5.5 0 0 0 0 1h5a.5.5 0 0 0 0-1h-5zm0 3a.5.5 0 0 0 0 1h3a.5.5 0 0 0 0-1h-3zm0 3a.5.5 0 0 0 0 1h1a.5.5 0 0 0 0-1h-1z"/>
-</svg>' :'<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-sort-down-alt" viewBox="0 0 16 16">
-  <path d="M3.5 3.5a.5.5 0 0 0-1 0v8.793l-1.146-1.147a.5.5 0 0 0-.708.708l2 1.999.007.007a.497.497 0 0 0 .7-.006l2-2a.5.5 0 0 0-.707-.708L3.5 12.293V3.5zm4 .5a.5.5 0 0 1 0-1h1a.5.5 0 0 1 0 1h-1zm0 3a.5.5 0 0 1 0-1h3a.5.5 0 0 1 0 1h-3zm0 3a.5.5 0 0 1 0-1h5a.5.5 0 0 1 0 1h-5zM7 12.5a.5.5 0 0 0 .5.5h7a.5.5 0 0 0 0-1h-7a.5.5 0 0 0-.5.5z"/>
-</svg>';
     }
 
+    /**
+     * @param $task_id
+     * delete task
+     */
     public function delete($task_id){
-        Tasks::where('id', $task_id)->delete();
+        if(Auth::user()->can('delete tasks')){
+            Tasks::where('id', $task_id)->delete();
+            $this->message_type = 'success';
+            $this->message =  __('Tasks deleted successfully');
+        }else{
+            $this->message_type = 'danger';
+            $this->message =  __('Not Allow!!!');
+        }
+    }
+
+    private function checkValidateJalali( $value, $fail){
+        $date = explode('-',$value);
+        if (count($date) != 3 && \Morilog\Jalali\CalendarUtils::checkDate($date[0], $date[1], $date[2], true)) {
+            $fail(__('validation.date_format_jalali'));
+        }
     }
 
     /**
@@ -54,59 +73,93 @@ class ShowTasks extends Component
      */
     public function render()
     {
+        //validate
+        $validator = Validator::make(
+            $this->search_tasks,
+            [
+                'name' => 'max:255|min:1',
+                'note' => '',
+                'status' => 'in:cancel,success,retarded,delete,doing,planned',
+                'date_start' => [
+                    'date_format:Y-m-d',
+                    'before_or_equal:date_end',
+                    function ($attribute, $value, $fail) {
+                        $this->checkValidateJalali( $value, $fail);
+                    },
+                ],
+                'date_end' => [
+                    'date_format:Y-m-d',
+                    'before_or_equal:date_end',
+                    function ($attribute, $value, $fail) {
+                        $this->checkValidateJalali( $value, $fail);
+                    },
+                ],
+                'time_start' => 'date_format:H:i:s|before_or_equal:time_end',
+                'time_end' => 'date_format:H:i:s|after_or_equal:time_start',
+                'user_id' => 'exists:users,id',
+                'create_by' => 'integer|exists:users,id',
+            ]
+        );
+
         //select tasks join user
-        $tasks = Tasks::join('users', 'users.id', '=', 'tasks.user_id')
-            ->join('users as creator', 'creator.id', '=', 'tasks.create_by_id')
-            ->select('tasks.*', 'users.name as user_name', 'users.last_name as user_last_name', 'creator.name as creator_name', 'creator.last_name as creator_last_name');
+        $tasks = Tasks::with(['user:id,name as user_name,last_name as user_last_name','creator:id,name as creator_name,last_name as creator_last_name']);
 
+        if ($validator->fails())
+        {
+            // The given data did not pass validation
+            $this->errors_message = $validator->errors()->all();
+        }else{
+            //reset error message not fail validator
+            $this->errors_message = [];
+        }
 
-        //check admin user show all list tasks
-        if(Auth::user()->is_admin != 1){
-            $tasks->where('users.id', Auth::user()->id);
+        //check permit view all tasks
+        if(!Auth::user()->can('view all tasks')){
+            $tasks->where('tasks.user_id', Auth::user()->id);
         }
 
         //filter name
-        if(isset($this->search_tasks['name'])){
+        if(isset($this->search_tasks['name']) && !$validator->errors()->has('name')){
             $tasks->where('tasks.name', 'like', '%'.$this->search_tasks['name'].'%');
         }
 
         //filter note
-        if(isset($this->search_tasks['note'])){
+        if(isset($this->search_tasks['note']) && !$validator->errors()->has('note')){
             $tasks->where('tasks.note', 'like', '%'.$this->search_tasks['note'].'%');
         }
 
         //filter status
-        if(isset($this->search_tasks['status']) && $this->search_tasks['status']){
+        if(isset($this->search_tasks['status']) && $this->search_tasks['status']  && !$validator->errors()->has('status')){
             $tasks->where('tasks.status', $this->search_tasks['status']);
         }
 
         //filter user_id
-        if(isset($this->search_tasks['user_id']) && $this->search_tasks['user_id']){
+        if(isset($this->search_tasks['user_id']) && $this->search_tasks['user_id']  && !$validator->errors()->has('user_id')){
             $tasks->where('tasks.user_id', $this->search_tasks['user_id']);
         }
 
         //filter create by id
-        if(isset($this->search_tasks['create_by']) && $this->search_tasks['create_by']){
+        if(isset($this->search_tasks['create_by']) && $this->search_tasks['create_by']  && !$validator->errors()->has('create_by')){
             $tasks->where('tasks.create_by_id', $this->search_tasks['create_by']);
         }
 
         //filter date start
-        if(isset($this->search_tasks['date_start']) && $this->search_tasks['date_start']){
+        if(isset($this->search_tasks['date_start']) && $this->search_tasks['date_start'] && !$validator->errors()->has('date_start')){
             $tasks->where('tasks.date' , '>=', $this->search_tasks['date_start']);
         }
 
         //filter date end
-        if(isset($this->search_tasks['date_end']) && $this->search_tasks['date_end']){
+        if(isset($this->search_tasks['date_end']) && $this->search_tasks['date_end'] && !$validator->errors()->has('date_end')){
             $tasks->where('tasks.date' , '<=', $this->search_tasks['date_end']);
         }
 
         //filter time start
-        if(isset($this->search_tasks['time_start']) && $this->search_tasks['time_start']){
+        if(isset($this->search_tasks['time_start']) && $this->search_tasks['time_start'] && !$validator->errors()->has('time_start')){
             $tasks->where('tasks.time' , '>=', $this->search_tasks['time_start']);
         }
 
         //filter time end
-        if(isset($this->search_tasks['time_end']) && $this->search_tasks['time_end']){
+        if(isset($this->search_tasks['time_end']) && $this->search_tasks['time_end'] && !$validator->errors()->has('time_end')){
             $tasks->where('tasks.time' , '<=', $this->search_tasks['time_end']);
         }
 
@@ -114,8 +167,7 @@ class ShowTasks extends Component
         $tasks=$tasks->orderBy($this->order_by, $this->order)->paginate(10);
 
         return view('livewire.show-tasks', [
-            'tasks' => $tasks,//tasks
-            'users' => User::all()//get list user for tasks user id & create user filter
+            'tasks' => $tasks//tasks
         ]);
     }
 }
